# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/promotion/mobile/{mobile}/stats": {
      "get": {
        "operationId": "PromotionController_getMobileStats",
        "parameters": [
          {
            "name": "mobile",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    },
...
    "/promotion/mobile/{mobile}/reset": {
      "post": {
        "operationId": "PromotionController_resetMobilePromotion",
        "parameters": [
          {
            "name": "mobile",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    },
...
    "/promotion/mobile/{mobile}/add": {
      "post": {
        "operationId": "PromotionController_addNewClient",
        "parameters": [
          {
            "name": "mobile",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    },
    "/promotion/mobile/{mobile}": {
      "delete": {
        "operationId": "PromotionController_removeClient",
        "parameters": [
          {
            "name": "mobile",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promotion"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS and Express API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}